[Test][TestSuite] Each problem is a group
