+++ test/util.ts
@@ -8,3 +8,14 @@ export function runAll(world:World, code:RawProgram) {
         world.runtime.step();
     }
 }
+
+
+function processWorld(xml: string): string[] {
+    const ans = xml.split(/\s+/g).filter(x => x);
+    return ans;
+}
+
+export function validateOutput(real: string, expected: string) {
+    expect(processWorld(real))
+        .toEqual(processWorld(expected));
+}
